### 2024-09-20 #7253 @editorialbot
[PRE REVIEW]: MIRTorch: A Differentiable Medical Image Reconstruction Toolbox
**Submitting author:** @GuanhuaW  (<a href="http://orcid.org/0000-0002-1622-5664">Guanhua Wang</a>)
**Repository:** https://github.com/guanhuaw/MIRTorch
**Branch with paper.md** (empty if default branch): feature/joss
**Version:** v0.1.2

Thanks for submitting your paper to JOSS @ GuanhuaW . **Currently, there isn't a JOSS editor assigned** to your paper.

@ GuanhuaW  if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
