### 2024-09-28 #7253 @ymzayek
Hi @guanhuaw, I am the handling editor for this submission. Can you please check the reference summary for the skipped and missing DOIs in https://github.com/openjournals/joss-reviews/issues/7253#issuecomment-2380587755 and update them or provide a reason for their omission?

Also let me know if you already have suggestions for reviewers. You can use the [potential reviewers list](https://reviewers.joss.theoj.org/lookup) provided at the top of this issue. You can also have a look at the authors in similar papers in JOSS https://github.com/openjournals/joss-reviews/issues/7253#issuecomment-2380588207.  If you have suggestions, please provide a list of potential reviewers here but do not use the @ character to avoid receiving unwanted notifications. Thank you in advance.
