### 2024-09-29 b2d1a84
update doi and contibuting guideline
