+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume = {49},
   number = {1},
   year = {1952},
-  publisher = {NBS Washington, DC}
+  publisher = {NBS Washington, DC},
+  doi = {10.6028/igres.049.044}
 }
 
 @article{deep_learning,
@@ -58,7 +59,8 @@
   number = {1},
   pages = {183--202},
   year = {2009},
-  publisher = {SIAM}
+  publisher = {SIAM},
+  doi = {10.1137/080716542}
 }
 
 @article{primal_dual,
@@ -68,7 +70,8 @@
   volume = {40},
   pages = {120--145},
   year = {2011},
-  publisher = {Springer}
+  publisher = {Springer},
+  doi = {10.1007/s10851-010-0251-1}
 }
 
 @article{proximal,
@@ -79,7 +82,8 @@
   number = {3},
   pages = {127--239},
   year = {2014},
-  publisher = {Now Publishers, Inc.}
+  publisher = {Now Publishers, Inc.},
+  doi = {10.1561/2400000003}
 }
 
 @article{sampling,
@@ -110,5 +114,6 @@
   number = {6},
   pages = {1768--1776},
   year = {2013},
-  publisher = {Wiley Online Library}
+  publisher = {Wiley Online Library},
+  doi = {10.1002/mrm.24389}
 }
