### 2024-10-03 #7253 @ymzayek
Hello @paquiteau, @jonbmartin and @H-Snoussi,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

MIRTorch: A Differentiable Medical Image Reconstruction Toolbox

You can find more information at the top of this Github issue.

Thank you in advance!
