### 2024-10-03 #7253 @paquiteau
Hello @ymzayek [^1]

Yes I can review it :)

@guanhuaw: shameless plug, but you can check out our work on Non-Cartesian MRI https://github.com/mind-inria/mri-nufft, that should a nice addition for your MRI linear operators (with performance improvements compared to torchkbnufft [^2])

[^1]: (nice to see you here btw !)
[^2]: https://github.com/mind-inria/mri-nufft-benchmark
