### 2024-10-09 #7340 @editorialbot
[REVIEW]: MIRTorch: A Differentiable Medical Image Reconstruction Toolbox
**Submitting author:** @ GuanhuaW  (<a href="http://orcid.org/0000-0002-1622-5664">Guanhua Wang</a>)
**Repository:** https://github.com/guanhuaw/MIRTorch
**Branch with paper.md** (empty if default branch): feature/joss
**Version:** v0.1.2
