### 2024-10-09 guanhuaw/MIRTorch#30 @paquiteau
[JOSS Review] Online Documentation and isntallation
Hello there,

I am reviewing MIRTorch for JOSS (https://github.com/openjournals/joss-reviews/issues/7340) After a brief look to the repo, I have seen some details that can be improve

- [ ] Is there an online documentation hosted somewhere (ideally with a gallery of the examples), everything seems to be setup in `docs` folder.
- [ ] I have seen that MIRTorch is available on PyPI: https://pypi.org/project/MIRTorch/, Is there any reason not to recommend its installation over the local clone and editable install recommended in the README ?

I will update this post I have more questions. MIRTorch seems to be good a software toolbox !
