### 2024-10-09 #7340 @paquiteau
Hello @GuanhuaW,
I have red the MIRTorch paper, and it seems that the state of the art regarding MRI Image reconstruction toolboxes There are quite a few and clearly stating how MIRTorch is different from others toolboxes (What does it do better/faster/easier ?). Without being exhaustive, I think [deepinv](https://github.com/deepinv/deepinv/) [^1], [pysap-mri](https://github.com/cea-cosmic/pysap-mri)[^1b] and  [scico](https://github.com/laln/scico) should be mentioned both are python toolboxes for solving Images Inverse problems, with deep learning capabilities, and MRI operators.

Regarding MRI capabilities, Our team maintains [mind-inra/mri-nufft](https://github.com/mind-inria/mri-nufft), which provides fast operators (at least faster than torchkbnufft, used by MIRTorch) for Non-Cartesian Imaging (and with autodiff capabilities as well, see our [examples](https://mind-inria.github.io/mri-nufft/generated/autoexamples/index.html)[^2]).

I have also opened an issue on your repo regarding documentation and installation instruction.

[^1]: Full Disclosure: Some of my coworkers at Inria are maintainer of deepinv
[^1b]: It is (not so actively) maintained in our team
[^2]: We are in the process of writing a JOSS paper as well
