### 2024-11-18 #7340 @jonbmartin
Hi @guanhuaw and @ymzayek, here is my initial review:

- Installation instructions: Would like to see required python version somewhere in the readme, docs page, etc. to facilitate casual check of python compatibility for potential users. Similarly, would also be nice to link to PyPi page somewhere in documentation and/or readme. (Aside: not a blocker for me w.r.t. review but would love to see installable through conda as well!)
- Installation/functionality: Very smooth and slick! Verified on local system w/ CPU and GPU w/ CUDA 11.8. All example code runs nicely. The only issue I encountered was with loading the .h5 file for the demo_3d.ipynb example - detailed in issue guanhuaw/MIRTorch#31. Not really MIRTorch-specific, just at FYI that I had trouble with that data. So not a block.
- Functionality documentation: This feels a little lacking. The code tutorials give excellent examples but there isn't much API documentation. I think that Frank Ong did a great job with this for SigPy as an example:  they 1) have pages outlining detailed usage for [specific functions](https://sigpy.readthedocs.io/en/latest/core_linop.html) and 2) some generic code usage examples like [this](https://sigpy.readthedocs.io/en/latest/guide_iter.html). I'd like to see a little more documentation developed before I check this box.
- Community guidelines: Also a little lacking. I know github has a standard approach but  the checklist asks for some specific guidelines, should probably add a bit more detail to satisfy.

Overall a huge fan! My critiques are really with documentation, the code itself seems solid by my judgement. Only 3 documentation boxes I haven't ticked.

Jon
